fix banker

+++ src/main/java/com/skyblock/skyblock/commands/menu/npc/BankerCommand.java
@@ -70,7 +70,7 @@ public class BankerCommand implements Command {
                 List<String> transactions = new ArrayList<>();
 
                 for (String transaction : recentTransactions) {
-                    int amount = Integer.parseInt(transaction.split(";")[0]);
+                    double amount = Double.parseDouble(transaction.split(";")[0]);
                     long time = Long.parseLong(transaction.split(";")[1]);
                     String by = transaction.split(";")[2];
 
@@ -79,7 +79,7 @@ public class BankerCommand implements Command {
                     String timeString = Util.calculateTimeAgoWithPeriodAndDuration(date.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime(), ZoneId.systemDefault());
                     String plusOrMinus = amount > 0 ? "&a+" : "&c-";
 
-                    transactions.add(plusOrMinus + " &6" + Util.formatInt(Util.assertPositive(amount)) + "&7, &e" + timeString + " &7by " + by);
+                    transactions.add(plusOrMinus + " &6" + Util.formatInt(Util.assertPositive((int) amount)) + "&7, &e" + timeString + " &7by " + by);
                 }
 
                 inventory.setItem(15, new ItemBuilder(ChatColor.GREEN + "Recent Transactions", Material.PAPER).setLore(
